# doc/redfish-openapi.yaml
openapi: 3.0.0
info:
  title: Redfish API
  version: 1.0.0
  description: Merged Redfish API specification
paths:
  /redfish/v1/:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceRoot_ServiceRoot'
          description: The response contains a representation of the ServiceRoot resource
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
...
  /redfish/v1/Systems:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComputerSystemCollection_ComputerSystemCollection'
          description: The response contains a representation of the ComputerSystemCollection
            resource
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
...
  /redfish/v1/Chassis:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChassisCollection_ChassisCollection'
          description: The response contains a representation of the ChassisCollection
            resource
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
...
  /redfish/v1/Managers:
    get:
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ManagerCollection_ManagerCollection'
          description: The response contains a representation of the ManagerCollection
            resource
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RedfishError'
          description: Error condition
...
components:
  schemas:
    ChassisCollection_ChassisCollection:
      additionalProperties: false
      description: The collection of `Chassis` resource instances.
      properties:
        '@odata.context':
          $ref: '#/components/schemas/odata-v4_context'
        '@odata.etag':
          $ref: '#/components/schemas/odata-v4_etag'
        '@odata.id':
          $ref: '#/components/schemas/odata-v4_id'
        '@odata.type':
          $ref: '#/components/schemas/odata-v4_type'
        Description:
          oneOf:
          - $ref: '#/components/schemas/Resource_Description'
          - enum:
            - null
          readOnly: true
        Members:
          description: The members of this collection.
          items:
            $ref: '#/components/schemas/odata-v4_idRef'
          readOnly: true
          type: array
          x-longDescription: This property shall contain an array of links to the
            members of this collection.
        Members@odata.count:
          $ref: '#/components/schemas/odata-v4_count'
        Members@odata.nextLink:
          $ref: '#/components/schemas/odata-v4_nextLink'
        Name:
          $ref: '#/components/schemas/Resource_Name'
          readOnly: true
        Oem:
          $ref: '#/components/schemas/Resource_Oem'
          description: The OEM extension property.
          x-longDescription: This property shall contain the OEM extensions.  All
            values for properties contained in this object shall conform to the Redfish
            Specification-described requirements.
      required:
      - Members
      - Members@odata.count
      - '@odata.id'
      - '@odata.type'
      - Name
      type: object
      x-longDescription: This resource shall represent a resource collection of `Chassis`
        instances for a Redfish implementation.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
...
    Resource_Description:
      description: The description of this resource.  Used for commonality in the
        schema definitions.
      type: string
      x-longDescription: This property shall contain the description of this resource.  The
        value shall conform with the 'Description' clause of the Redfish Specification.
    Resource_Health:
      enum:
      - OK
      - Warning
      - Critical
      type: string
      x-enumDescriptions:
        Critical: A critical condition requires immediate attention.
        OK: Normal.
        Warning: A condition requires attention.
    Resource_Id:
      description: The unique identifier for this resource within the collection of
        similar resources.
      type: string
      x-longDescription: This property shall contain the identifier for this resource.  The
        value shall conform with the 'Id' clause of the Redfish Specification.
...
    Resource_Name:
      description: The name of the resource or array member.
      type: string
      x-longDescription: This property shall contain the name of this resource or
        array member.  The value shall conform with the 'Name' clause of the Redfish
        Specification.
    Resource_Oem:
      additionalProperties: true
      description: The OEM extension.
      properties: {}
      type: object
      x-longDescription: This object represents the OEM properties.  The resource
        values shall comply with the Redfish Specification-described requirements.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
        ^[A-Za-z0-9_]+$:
          $ref: '#/components/schemas/Resource_OemObject'
    Resource_OemObject:
      additionalProperties: true
      description: The base type for an OEM extension.
      properties: {}
      type: object
      x-longDescription: This object represents the base type for an OEM object.  The
        resource values shall comply with the Redfish Specification-described requirements.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
...
    Message_Message:
      additionalProperties: false
      description: The message that the Redfish service returns.
      properties:
        Message:
          description: The human-readable message.
          readOnly: true
          type: string
          x-longDescription: This property shall contain a human-readable message.
        MessageArgs:
          description: An array of message arguments that are substituted for the
            arguments in the message when looked up in the message registry.
          items:
            type: string
          readOnly: true
          type: array
          x-longDescription: This property shall contain an array of message arguments
            that are substituted for the arguments in the message when looked up in
            the message registry.  It has the same semantics as the `MessageArgs`
            property in the Redfish `MessageRegistry` schema.  If the corresponding
            `ParamType` value contains `number`, the service shall convert the number
            to a string representation of the number.
        MessageId:
          description: The identifier for the message.
          readOnly: true
          type: string
          x-longDescription: This property shall contain a `MessageId`, as defined
            in the 'MessageId format' clause of the Redfish Specification.
        MessageSeverity:
          $ref: '#/components/schemas/Resource_Health'
          description: The severity of the message.
          readOnly: true
          x-longDescription: This property shall contain the severity of the message.  Services
            can replace the value defined in the message registry with a value more
            applicable to the implementation.
          x-versionAdded: v1_1_0
        Oem:
          $ref: '#/components/schemas/Resource_Oem'
          description: The OEM extension property.
          x-longDescription: This property shall contain the OEM extensions.  All
            values for properties contained in this object shall conform to the Redfish
            Specification-described requirements.
        RelatedProperties:
          description: A set of properties described by the message.
          items:
            type: string
          readOnly: true
          type: array
          x-longDescription: This property shall contain an array of RFC6901-defined
            JSON pointers indicating the properties described by the message, if appropriate
            for the message.
        Resolution:
          description: Used to provide suggestions on how to resolve the situation
            that caused the message.
          readOnly: true
          type: string
          x-longDescription: This property shall contain the resolution of the message.  Services
            can replace the resolution defined in the message registry with a more
            specific resolution in message payloads.
        ResolutionSteps:
          description: The list of recommended steps to resolve the situation that
            caused the message.
          items:
            $ref: '#/components/schemas/ResolutionStep_ResolutionStep'
          type: array
          x-longDescription: This property shall contain an array of recommended steps
            to resolve the situation that caused the message.  This property shall
            not be present if the `MessageSeverity` or `Severity` properties contain
            `OK`.
          x-versionAdded: v1_2_0
        Severity:
          deprecated: true
          description: The severity of the message.
          readOnly: true
          type: string
          x-deprecatedReason: This property has been deprecated in favor of `MessageSeverity`,
            which ties the values to the enumerations defined for the `Health` property
            within `Status`.
          x-longDescription: This property shall contain the severity of the message,
            as defined in the 'Status' clause of the Redfish Specification.  Services
            can replace the value defined in the message registry with a value more
            applicable to the implementation.
          x-versionDeprecated: v1_1_0
      required:
      - MessageId
      type: object
      x-longDescription: This type shall contain a message that the Redfish service
        returns, as described in the Redfish Specification.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
    ServiceRoot_ServiceRoot:
      additionalProperties: false
      description: The `ServiceRoot` schema describes the root of the Redfish service,
        located at the '/redfish/v1' URI.  All other resources accessible through
        the Redfish interface on this device are linked directly or indirectly from
        the service root.
      properties:
        '@odata.context':
          $ref: '#/components/schemas/odata-v4_context'
        '@odata.id':
          $ref: '#/components/schemas/odata-v4_id'
        '@odata.type':
          $ref: '#/components/schemas/odata-v4_type'
        Chassis:
          $ref: '#/components/schemas/odata-v4_idRef'
          description: The link to a collection of chassis.
          readOnly: true
          x-longDescription: This property shall contain a link to a resource collection
            of type `ChassisCollection`.
        Description:
          oneOf:
          - $ref: '#/components/schemas/Resource_Description'
          - enum:
            - null
          readOnly: true
        Id:
          $ref: '#/components/schemas/Resource_Id'
          readOnly: true
        Managers:
          $ref: '#/components/schemas/odata-v4_idRef'
          description: The link to a collection of managers.
          readOnly: true
          x-longDescription: This property shall contain a link to a resource collection
            of type `ManagerCollection`.
        Name:
          $ref: '#/components/schemas/Resource_Name'
          readOnly: true
        RedfishVersion:
          description: The version of the Redfish service.
          pattern: ^\d+\.\d+\.\d+$
          readOnly: true
          type: string
          x-longDescription: This property shall represent the Redfish protocol version,
            as specified in the 'Protocol version' clause of the Redfish Specification,
            to which this service conforms.
        Systems:
          $ref: '#/components/schemas/odata-v4_idRef'
          description: The link to a collection of systems.
          readOnly: true
          x-longDescription: This property shall contain a link to a resource collection
            of type `ComputerSystemCollection`.
        UUID:
          description: Unique identifier for a service instance.  When SSDP is used,
            this value contains the same UUID returned in an HTTP `200 OK` response
            from an SSDP `M-SEARCH` request during discovery.
          nullable: true
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
          readOnly: true
          type: string
          x-longDescription: This property shall contain the identifier of the Redfish
            service instance.  If SSDP is used, this value shall contain the same
            UUID returned in an HTTP `200 OK` response from an SSDP `M-SEARCH` request
            during discovery.  RFC4122 describes methods to use to create a UUID value.  The
            value should be considered to be opaque.  Client software should only
            treat the overall value as a universally unique identifier and should
            not interpret any subfields within the UUID.
      required:
      - '@odata.id'
      - '@odata.type'
      - Id
      - Name
      type: object
      x-longDescription: This resource shall represent the root of the Redfish service.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
    odata-v4_context:
      description: The OData description of a payload.
      format: uri-reference
      readOnly: true
      type: string
      x-longDescription: The value of this property shall be the context URL that
        describes the resource according to OData-Protocol and shall be of the form
        defined in the Redfish specification.
    odata-v4_count:
      description: The number of items in a collection.
      format: int64
      readOnly: true
      type: integer
      x-longDescription: The value of this property shall be an integer representing
        the number of items in a collection.
    odata-v4_etag:
      description: The current ETag of the resource.
      readOnly: true
      type: string
      x-longDescription: The value of this property shall be a string that is defined
        by the ETag HTTP header definition in RFC7232.
    odata-v4_id:
      description: The unique identifier for a resource.
      format: uri-reference
      readOnly: true
      type: string
      x-longDescription: The value of this property shall be the unique identifier
        for the resource and it shall be of the form defined in the Redfish specification.
    odata-v4_idRef:
      additionalProperties: false
      description: A reference to a resource.
      properties:
        '@odata.id':
          $ref: '#/components/schemas/odata-v4_id'
      type: object
      x-longDescription: The value of this property shall be used for references to
        a resource.
    odata-v4_nextLink:
      description: The URI to the resource containing the next set of partial members.
      format: uri-reference
      readOnly: true
      type: string
      x-longDescription: The value of this property shall be a URI to a resource,
        with the same @odata.type, containing the next set of partial members.
    odata-v4_type:
      description: The type of a resource.
      readOnly: true
      type: string
      x-longDescription: The value of this property shall be a URI fragment that specifies
        the type of the resource and it shall be of the form defined in the Redfish
        specification.
    ComputerSystemCollection_ComputerSystemCollection:
      additionalProperties: false
      description: The collection of `ComputerSystem` resource instances.
      properties:
        '@odata.context':
          $ref: '#/components/schemas/odata-v4_context'
        '@odata.etag':
          $ref: '#/components/schemas/odata-v4_etag'
        '@odata.id':
          $ref: '#/components/schemas/odata-v4_id'
        '@odata.type':
          $ref: '#/components/schemas/odata-v4_type'
        Description:
          oneOf:
          - $ref: '#/components/schemas/Resource_Description'
          - enum:
            - null
          readOnly: true
        Members:
          description: The members of this collection.
          items:
            $ref: '#/components/schemas/odata-v4_idRef'
          readOnly: true
          type: array
          x-longDescription: This property shall contain an array of links to the
            members of this collection.
        Members@odata.count:
          $ref: '#/components/schemas/odata-v4_count'
        Members@odata.nextLink:
          $ref: '#/components/schemas/odata-v4_nextLink'
        Name:
          $ref: '#/components/schemas/Resource_Name'
          readOnly: true
        Oem:
          $ref: '#/components/schemas/Resource_Oem'
          description: The OEM extension property.
          x-longDescription: This property shall contain the OEM extensions.  All
            values for properties contained in this object shall conform to the Redfish
            Specification-described requirements.
      required:
      - Members
      - Members@odata.count
      - '@odata.id'
      - '@odata.type'
      - Name
      type: object
      x-longDescription: This resource shall represent a resource collection of `ComputerSystem`
        instances for a Redfish implementation.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
...
    ManagerCollection_ManagerCollection:
      additionalProperties: false
      description: The collection of `Manager` resource instances.
      properties:
        '@odata.context':
          $ref: '#/components/schemas/odata-v4_context'
        '@odata.etag':
          $ref: '#/components/schemas/odata-v4_etag'
        '@odata.id':
          $ref: '#/components/schemas/odata-v4_id'
        '@odata.type':
          $ref: '#/components/schemas/odata-v4_type'
        Description:
          oneOf:
          - $ref: '#/components/schemas/Resource_Description'
          - enum:
            - null
          readOnly: true
        Members:
          description: The members of this collection.
          items:
            $ref: '#/components/schemas/odata-v4_idRef'
          readOnly: true
          type: array
          x-longDescription: This property shall contain an array of links to the
            members of this collection.
        Members@odata.count:
          $ref: '#/components/schemas/odata-v4_count'
        Members@odata.nextLink:
          $ref: '#/components/schemas/odata-v4_nextLink'
        Name:
          $ref: '#/components/schemas/Resource_Name'
          readOnly: true
        Oem:
          $ref: '#/components/schemas/Resource_Oem'
          description: The OEM extension property.
          x-longDescription: This property shall contain the OEM extensions.  All
            values for properties contained in this object shall conform to the Redfish
            Specification-described requirements.
      required:
      - Members
      - Members@odata.count
      - '@odata.id'
      - '@odata.type'
      - Name
      type: object
      x-longDescription: This resource shall represent a resource collection of `Manager`
        instances for a Redfish implementation.
      x-patternProperties:
        ^([a-zA-Z_][a-zA-Z0-9_]*)?@(odata|Redfish|Message)\.[a-zA-Z_][a-zA-Z0-9_]*$:
          description: This property shall specify a valid odata or Redfish property.
...
    RedfishError:
      description: The error payload from a Redfish service.
      properties:
        error:
          description: The properties that describe an error from a Redfish service.
          properties:
            '@Message.ExtendedInfo':
              description: An array of messages describing one or more error messages.
              items:
                $ref: '#/components/schemas/Message_Message'
              type: array
              x-longDescription: This property shall be an array of message objects
                describing one or more error messages.
            code:
              description: A string indicating a specific `MessageId` from a message
                registry.
              readOnly: true
              type: string
              x-longDescription: This property shall contain a string indicating a
                specific `MessageId` from a message registry.
            message:
              description: A human-readable error message corresponding to the message
                in a message registry.
              readOnly: true
              type: string
              x-longDescription: This property shall contain a human-readable error
                message corresponding to the message in a message registry.
          required:
          - code
          - message
          type: object
          x-longDescription: The Redfish Specification-described type shall contain
            properties that describe an error from a Redfish service.
      required:
      - error
      type: object
      x-longDescription: The Redfish Specification-described type shall contain an
        error payload from a Redfish service.